add condition if executable doesn't exist

+++ run.sh
@@ -1,9 +1,11 @@
-echo "Compiling executable binary..."
-cd src
-make compile
-cp gol ../
-cd ..
-echo "Compiled executable \"gol\"."
+if [ ! -e gol ]; then
+    echo "Compiling executable binary..."
+    cd src
+    make compile
+    cp gol ../
+    cd ..
+    echo "Compiled executable \"gol\"."
+fi
 
 if [ $# -lt 1 ]; then
     ./gol < samples/gosper_gun.txt
